Shared between UWP & NET code is encapsulated into separate project.
