+++ windows/CodePush.Shared/CodePushConstants.cs
