+++ windows/CodePush.Shared/CodePushReactPackage.cs
@@ -1,25 +1,21 @@
 ﻿using Newtonsoft.Json.Linq;
-using Newtonsoft.Json;
 using ReactNative;
 using ReactNative.Bridge;
 using ReactNative.Modules.Core;
 using ReactNative.UIManager;
 using System;
 using System.Collections.Generic;
-using System.Reflection;
 using System.Threading.Tasks;
-using Windows.ApplicationModel;
-using Windows.Storage;
-using Windows.Web.Http;
+
 
 namespace CodePush.ReactNative
 {
     public sealed class CodePushReactPackage : IReactPackage
     {
         private static CodePushReactPackage CurrentInstance;
-        private static bool NeedToReportRollback = false;
+        //private static bool NeedToReportRollback = false;
 
-        private CodePushNativeModule _codePushNativeModule;
+        //private CodePushNativeModule _codePushNativeModule;
 
         internal string AppVersion { get; private set; }
         internal string DeploymentKey { get; private set; }
@@ -31,7 +27,7 @@ namespace CodePush.ReactNative
 
         public CodePushReactPackage(string deploymentKey, ReactPage mainPage)
         {
-            AppVersion = Package.Current.Id.Version.Major + "." + Package.Current.Id.Version.Minor + "." + Package.Current.Id.Version.Build;
+            AppVersion = CodePushUtils.GetAppVersion();
             DeploymentKey = deploymentKey;
             MainPage = mainPage;
             UpdateManager = new UpdateManager();
@@ -85,7 +81,7 @@ namespace CodePush.ReactNative
         {
             AssetsBundleFileName = assetsBundleFileName;
             string binaryJsBundleUrl = CodePushConstants.AssetsBundlePrefix + assetsBundleFileName;
-            var binaryResourcesModifiedTime = await GetBinaryResourcesModifiedTimeAsync().ConfigureAwait(false);
+            var binaryResourcesModifiedTime = await FileUtils.GetBinaryResourcesModifiedTimeAsync(AssetsBundleFileName).ConfigureAwait(false);
             var packageFile = await UpdateManager.GetCurrentPackageBundleAsync(AssetsBundleFileName).ConfigureAwait(false);
             if (packageFile == null)
             {
@@ -111,7 +107,7 @@ namespace CodePush.ReactNative
             {
                 CodePushUtils.LogBundleUrl(packageFile.Path);
                 IsRunningBinaryVersion = false;
-                return CodePushConstants.FileBundlePrefix + packageFile.Path.Replace(ApplicationData.Current.LocalFolder.Path, "").Replace("\\", "/");
+                return CodePushConstants.FileBundlePrefix + packageFile.Path.Replace(CodePushUtils.GetAppFolder(), "").Replace("\\", "/");
             }
             else
             {
@@ -132,13 +128,6 @@ namespace CodePush.ReactNative
 
         #region Internal methods
 
-        internal async Task<long> GetBinaryResourcesModifiedTimeAsync()
-        {
-            var assetJSBundleFile = await StorageFile.GetFileFromApplicationUriAsync(new Uri(CodePushConstants.AssetsBundlePrefix + AssetsBundleFileName)).AsTask().ConfigureAwait(false);
-            var fileProperties = await assetJSBundleFile.GetBasicPropertiesAsync().AsTask().ConfigureAwait(false);
-            return fileProperties.DateModified.ToUnixTimeMilliseconds();
-        }
-        
         internal void InitializeUpdateAfterRestart()
         {
             JObject pendingUpdate = SettingsManager.GetPendingUpdate();
@@ -150,7 +139,7 @@ namespace CodePush.ReactNative
                     // Pending update was initialized, but notifyApplicationReady was not called.
                     // Therefore, deduce that it is a broken update and rollback.
                     CodePushUtils.Log("Update did not finish loading the last time, rolling back to a previous version.");
-                    NeedToReportRollback = true;
+                    //NeedToReportRollback = true;
                     RollbackPackageAsync().Wait();
                 }
                 else
@@ -159,7 +148,7 @@ namespace CodePush.ReactNative
                     // Clear the React dev bundle cache so that new updates can be loaded.
                     if (MainPage.UseDeveloperSupport)
                     {
-                        ClearReactDevBundleCacheAsync().Wait();
+                        FileUtils.ClearReactDevBundleCacheAsync().Wait();
                     }
                     // Mark that we tried to initialize the new update, so that if it crashes,
                     // we will know that we need to rollback when the app next starts.
@@ -178,15 +167,6 @@ namespace CodePush.ReactNative
         #endregion
 
         #region Private methods
-
-        private async Task ClearReactDevBundleCacheAsync()
-        {
-            var devBundleCacheFile = (StorageFile)await ApplicationData.Current.LocalFolder.TryGetItemAsync(CodePushConstants.ReactDevBundleCacheFileName).AsTask().ConfigureAwait(false);
-            if (devBundleCacheFile != null)
-            {
-                await devBundleCacheFile.DeleteAsync().AsTask().ConfigureAwait(false);
-            }
-        }
 
         private async Task RollbackPackageAsync()
         {
