+++ windows/CodePush.Shared/InstallMode.cs
