+++ windows/CodePush.Shared/MinimumBackgroundListener.cs
