+++ windows/CodePush.Shared/UpdateState.cs
