+++ windows/CodePush/CodePush.csproj
@@ -107,13 +107,8 @@
     <None Include="project.json" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="CodePushReactPackage.cs" />
-    <Compile Include="CodePushConstants.cs" />
-    <Compile Include="MinimumBackgroundListener.cs" />
-    <Compile Include="InstallMode.cs" />
-    <Compile Include="CodePushNativeModule.cs" />
+    <Compile Include="CodePushNativeModuleImpl.cs" />
     <Compile Include="UpdateManager.cs" />
-    <Compile Include="UpdateState.cs" />
     <Compile Include="UpdateUtils.cs" />
     <Compile Include="CodePushUtils.cs" />
     <Compile Include="FileUtils.cs" />
@@ -127,11 +122,12 @@
     </SDKReference>
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\..\react-native-windows\ReactWindows\ReactNative\ReactNative.csproj">
+    <ProjectReference Include="..\..\Examples\CodePushDemoApp\node_modules\react-native-windows\ReactWindows\ReactNative\ReactNative.csproj">
       <Project>{c7673ad5-e3aa-468c-a5fd-fa38154e205c}</Project>
       <Name>ReactNative</Name>
     </ProjectReference>
   </ItemGroup>
+  <Import Project="..\CodePush.Shared\CodePush.Shared.projitems" Label="Shared" />
   <PropertyGroup Condition=" '$(VisualStudioVersion)' == '' or '$(VisualStudioVersion)' &lt; '14.0' ">
     <VisualStudioVersion>14.0</VisualStudioVersion>
   </PropertyGroup>
