+++ windows/CodePush/Properties/AssemblyInfo.cs
