+++ windows/CodePush/Properties/CodePush.rd.xml
