+++ windows/CodePush/SettingsManager.cs
