+++ windows/CodePush/UpdateManager.cs
