+++ windows/CodePush/UpdateUtils.cs
