+++ windows/CodePush/project.json
